+++ lib/spectre/diagnostic.rb
@@ -21,6 +21,14 @@ module Spectre
         def duration
           @@end_time - @@start_time
         end
+
+        def started_at
+          @@start_time
+        end
+
+        def finished_at
+          @@end_time
+        end
       end
 
       Spectre.delegate :start_watch, :stop_watch, :duration, :measure, to: self