+++ CHANGELOG.md
@@ -4,6 +4,7 @@
  - New mixin methods for defining required and optional parameters added
  - Result logging for `observe` added
  - HTTP `read_timeout` made configurable via `timeout` method and property in yml file and default value set to 180s
+ - Added new functions `started_at` and `finished_at` to `spectre/diagnostic`
 
 
 ### v1.11.0